requirements.txt
Django>=5.0
psycopg2-binary
akshare
pandas
numpy
scipy
scikit-learn
celery[redis]>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
python-dotenv
django-unfold>=0.30.0
# TA-Lib  # Requires separate installation: conda install -c conda-forge ta-lib
